fix(checker): split symbol factory identity helper
