+++ crates/tsz-checker/src/types/computation/helpers.rs
@@ -901,51 +901,6 @@ impl<'a> CheckerState<'a> {
         annotation_type
     }
 
-    /// Upgrade the value type of an unannotated `const X = Symbol()` /
-    /// `const X = Symbol.for(...)` declaration to its `unique symbol` value
-    /// identity (`UniqueSymbol(SymbolRef(X))`), keyed on the variable's own
-    /// binder symbol.
-    ///
-    /// TypeScript gives an unannotated `const` initialized with a global
-    /// `Symbol(...)`/`Symbol.for(...)` factory call the distinct value identity
-    /// `typeof X` (a `unique symbol`), not the general `symbol` type the call
-    /// signature returns. The annotation form (`const X: unique symbol`) is
-    /// centralized in `const_unique_symbol_value_type`; this is the sibling
-    /// upgrade for the initializer form, so every value-typing path — direct
-    /// variable typing, the merged `type X = typeof X` value path, and
-    /// cross-arena value-declaration delegation — agrees on the same identity.
-    /// Without it, a `Symbol.for` const that is name-merged with a same-named
-    /// type alias loses its identity and degrades to wide `symbol`, so an
-    /// object literal `{ [X]: v }` synthesizes a `[k: symbol]: V` index
-    /// signature instead of the symbol-keyed member the interface expects.
-    ///
-    /// Returns `None` when the declaration is not a `const` whose initializer is
-    /// a verified global `Symbol`/`Symbol.for` factory call.
-    pub(crate) fn const_symbol_factory_unique_value_type(
-        &self,
-        decl_idx: NodeIndex,
-    ) -> Option<TypeId> {
-        let var_decl = self
-            .ctx
-            .arena
-            .get(decl_idx)
-            .and_then(|node| self.ctx.arena.get_variable_declaration(node))?;
-        if var_decl.initializer.is_none() || !self.is_const_variable_declaration(decl_idx) {
-            return None;
-        }
-        if !(self.is_symbol_call_initializer(var_decl.initializer)
-            || self.is_symbol_for_call_initializer(var_decl.initializer))
-        {
-            return None;
-        }
-        let sym_id = self.get_symbol_id_for_variable_name(var_decl.name)?;
-        Some(
-            self.ctx
-                .types
-                .unique_symbol(tsz_solver::SymbolRef(sym_id.0)),
-        )
-    }
-
     /// Get type of variable declaration.
     ///
     /// Computes the type of variable declarations like `let x: number = 5` or `const y = "hello"`.
@@ -1013,20 +968,6 @@ impl<'a> CheckerState<'a> {
                     widened,
                     self.ctx.strict_null_checks(),
                 );
-            }
-
-            // `const k = Symbol()` / `const k = Symbol.for(...)` — infer unique
-            // symbol type. In TypeScript, unannotated const declarations
-            // initialized with global symbol factory calls get a unique symbol
-            // type (typeof k), not the general `symbol` type.
-            if (self.is_symbol_call_initializer(var_decl.initializer)
-                || self.is_symbol_for_call_initializer(var_decl.initializer))
-                && let Some(sym_id) = self.get_symbol_id_for_variable_name(var_decl.name)
-            {
-                return self
-                    .ctx
-                    .types
-                    .unique_symbol(tsz_solver::SymbolRef(sym_id.0));
             }
 
             // const: preserve literal type from the initializer directly.
