+++ crates/tsz-checker/src/types/computation/mod.rs
@@ -59,5 +59,6 @@ pub(crate) mod object_literal_circularity;
 pub(crate) mod object_literal_context;
 pub(crate) mod object_literal_support;
 pub(crate) mod object_literal_widening;
+mod symbol_factory_identity;
 pub(crate) mod tagged_template;
 pub(crate) mod type_operators;
